refactor: rename GoogleSQL to Spanner
