+++ .sample.env
@@ -4,6 +4,5 @@ SQLITE_CONN_STRING="testdb.sqlite3"
 SQLSERVER_CONN_STRING="server=localhost\SQLExpress;database=testfixtures_test;user id=sa;password=sqlserver;encrypt=disable"
 CRDB_CONN_STRING="host=cockroachdb user=root dbname=defaultdb port=26257 sslmode=disable"
 CLICKHOUSE_CONN_STRING="clickhouse://clickhouse:clickhouse@clickhouse:9000/testfixtures_test?debug=false"
-# TODO: fill the value
-GOOGLESQL_CONN_STRING="projects/test-project/instances/test-instance/databases/testdb"
-SPANNER_EMULATOR_HOST="googlesql:9010"
+SPANNER_CONN_STRING="projects/test-project/instances/test-instance/databases/testdb"
+SPANNER_EMULATOR_HOST="spanner:9010"
