+++ Taskfile.yml
@@ -52,11 +52,11 @@ tasks:
       - task: test-db
         vars: {DATABASE: clickhouse}
 
-  test:googlesql:
+  test:spanner:
     desc: Test Spanner with GoogleSQL dialect
     cmds:
       - task: test-db
-        vars: {DATABASE: googlesql}
+        vars: {DATABASE: spanner}
 
   test-db:
     cmds:
@@ -79,4 +79,4 @@ tasks:
   docker:test:
     cmds:
       - docker compose down -v --remove-orphans
-      - docker compose run testfixtures go test -v -tags 'postgresql sqlite mysql sqlserver cockroachdb clickhouse googlesql'
+      - docker compose run testfixtures go test -v -tags 'postgresql sqlite mysql sqlserver cockroachdb clickhouse spanner'
