+++ docker-compose.yml
@@ -9,7 +9,7 @@ services:
       - sqlserver
       - cockroachdb
       - clickhouse
-      - googlesql
+      - spanner
     environment:
       PGPASSWORD: postgres
       PG_CONN_STRING: host=postgresql user=postgres dbname=testfixtures_test port=5432 sslmode=disable
@@ -18,8 +18,8 @@ services:
       SQLSERVER_CONN_STRING: server=sqlserver;database=master;user id=sa;password=SQL@1server;encrypt=disable
       CRDB_CONN_STRING: host=cockroachdb user=root dbname=defaultdb port=26257 sslmode=disable
       CLICKHOUSE_CONN_STRING: clickhouse://clickhouse:clickhouse@clickhouse:9000/testfixtures_test?debug=false
-      GOOGLESQL_CONN_STRING: projects/test-project/instances/test-instance/databases/testdb
-      SPANNER_EMULATOR_HOST: googlesql:9010
+      SPANNER_CONN_STRING: projects/test-project/instances/test-instance/databases/testdb
+      SPANNER_EMULATOR_HOST: spanner:9010
 
   postgresql:
     image: postgres:16.3-alpine
@@ -54,7 +54,7 @@ services:
       CLICKHOUSE_USER: clickhouse
       CLICKHOUSE_PASSWORD: clickhouse
 
-  googlesql:
+  spanner:
     image: gcr.io/cloud-spanner-emulator/emulator
     ports:
       - "19010:9010"
