+++ helper.go
@@ -46,7 +46,7 @@ type batchSplitter interface { //nolint
 
 var (
 	_ helper = &clickhouse{}
-	_ helper = &googleSQL{}
+	_ helper = &spanner{}
 	_ helper = &mySQL{}
 	_ helper = &postgreSQL{}
 	_ helper = &sqlite{}
