+++ testdata/schema/spanner.sql
