+++ testfixtures.go
@@ -112,7 +112,7 @@ func Database(db *sql.DB) func(*Loader) error {
 // Dialect informs Loader about which database dialect you're using.
 //
 // Possible options are "postgresql", "timescaledb", "mysql", "mariadb",
-// "sqlite", "sqlserver", "clickhouse", "googlesql".
+// "sqlite", "sqlserver", "clickhouse", "spanner".
 func Dialect(dialect string) func(*Loader) error {
 	return func(l *Loader) error {
 		h, err := helperForDialect(dialect)
@@ -137,7 +137,7 @@ func helperForDialect(dialect string) (helper, error) {
 	case "clickhouse":
 		return &clickhouse{}, nil
 	case "spanner":
-		return &googleSQL{}, nil // TODO: return the impl
+		return &spanner{}, nil // TODO: return the impl
 	default:
 		return nil, fmt.Errorf(`testfixtures: unrecognized dialect "%s"`, dialect)
 	}
